+++ src/project/lib/structs/repository_info/new.rs
@@ -1,11 +1,17 @@
+use std::path::Path;
 use crate::RepositoryInfo;
+use crate::utils::helpers::Result;
 use git2::Repository;
-use std::error::Error;
 
 impl RepositoryInfo {
-    pub fn new(repo: Repository) -> Result<Self, Box<dyn Error>> {
-        let name = repo
-            .path()
+    pub fn new<P>(path: P) -> Result<Self>
+    where
+        P: AsRef<Path>,
+    {
+        let path = path.as_ref();
+        
+        let repo = Repository::open(&path)?;
+        let name = path
             .parent()
             .and_then(|p| p.file_name())
             .and_then(|n| n.to_str())